requirements.txt
###### Requirements without Version Specifiers ######`

###### Requirements with Version Specifiers ######`
paho-mqtt >= 1.5.1             # Version must be at least version 1.5.1
speedtest-cli >= 1.0.0
python-dotenv >= 0.10.1
